pyproject.toml
[tool.poetry]
name = "transaction-analysis"
version = "0.1.0"
description = ""
authors = ["Serg Kuchukov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
python-dateutil = "^2.9.0.post0"
freezegun = "^1.5.1"


[tool.poetry.group.develop.dependencies]
mypy = "^1.11.2"
black = "^24.10.0"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
